+++ controllers/journalEntryController.js
@@ -1,10 +1,15 @@
 const JournalEntry = require('../models/JournalEntry');
 const JournalEntryDetail = require('../models/JournalEntryDetail');
+const AccountingPeriod = require('../models/AccountingPeriods');
 
 const createJournalEntry = async (req, res) => {
     try {
         const { body, tenant } = req;
-        const { date, reference, description, details } = body;
+        const { date, description, details } = body;
+
+        if (!date || !description) {
+            return res.status(400).json({ status: "error", message: "Fecha y descripción son requeridos", data: null });
+        }
 
         if (!details || details.length === 0) {
             return res.status(400).json({ status: "error", message: "Los detalles no pueden estar vacíos", data: null });
@@ -18,29 +23,29 @@ const createJournalEntry = async (req, res) => {
             return res.status(400).json({ status: "error", message: "El débito y el crédito deben ser iguales", data: null });
         }
 
-        // valida referencias unicas
-        const existReference = await JournalEntry.findOne({ reference, tenant });
+        // Obtiene el periodo contable abierto por rango de fechas
+        const currentPeriod = await AccountingPeriod.findOne({ tenant, status: 'open', start_date: { $lte: date }, end_date: { $gte: date } });
 
-        if (existReference) {
-            return res.status(400).json({ status: "error", message: `Ya existe un registro con la referencia ${reference}`, data: null });
+        if (!currentPeriod) {
+            return res.status(400).json({ status: "error", message: "El periodo contable no existe o esta cerrado", data: null });
         }
 
         // Crear y guardar la cabecera
         const journalEntry = await JournalEntry.create({
             tenant,
             date,
-            reference,
-            description
+            description,
+            period: currentPeriod._id
         });
 
         // Preparar los detalles del asiento contable
-        const journalEntryDetails = details.map(({ ledgerAccount, debit, credit, costCenter }) => ({
+        const journalEntryDetails = details.map(({ account, debit, credit, cost_center }) => ({
             tenant,
-            ledgerAccount,
+            account,
             debit: debit || 0,
             credit: credit || 0,
-            costCenter: costCenter || null,
-            journalEntry: journalEntry._id,
+            cost_center: cost_center || null,
+            journal_entry: journalEntry._id,
         }));
 
         // Insertar los detalles en paralelo con `insertMany`
@@ -85,8 +90,9 @@ const getJournalEntries = async (req, res) => {
 
         const journalEntries = await JournalEntry.find({ tenant }).populate({
             path: 'details',
-            populate: { path: 'ledgerAccount costCenter' }
+            populate: { path: 'account cost_center' }
         });
+
         res.status(201).json({ status: 'ok', message: 'Journal Entries found successfully', data: journalEntries });
     } catch (err) {
         res.status(400).json({ status: 'error', message: `Error al encontrar los asientos contables: ${err.message}`, data: null });
@@ -97,10 +103,20 @@ const updateJournalEntry = async (req, res) => {
     try {
         const { _id } = req.params;
         const { body, tenant } = req;
-        const { date, reference, description, details } = body;
+        const { date, description, details } = body;
+
+        if (!date || !description) {
+            return res.status(400).json({ status: "error", message: "Fecha y descripción son requeridos", data: null });
+        }
 
         if (!details || details.length === 0) {
             return res.status(400).json({ status: "error", message: "Detalle no puede ser vacio", data: null });
+        }
+
+        // Obtiene el periodo contable abierto por rango de fechas
+        const currentPeriod = await AccountingPeriod.findOne({ tenant, status: 'open', start_date: { $lte: date }, end_date: { $gte: date } });
+        if (!currentPeriod) {
+            return res.status(400).json({ status: "error", message: "El periodo contable no existe o esta cerrado", data: null });
         }
 
         // Valida que el asiento este cuadrado
@@ -114,21 +130,21 @@ const updateJournalEntry = async (req, res) => {
         // Actualizar la cabecera
         const journalEntry = await JournalEntry.findByIdAndUpdate(
             { _id, tenant },
-            { date, reference, description },
+            { date, description, period: currentPeriod._id },
             { new: true }
         );
 
         // Eliminar los detalles existentes
-        await JournalEntryDetail.deleteMany({ journalEntry: journalEntry._id });
+        await JournalEntryDetail.deleteMany({ journal_entry: journalEntry._id });
 
         // Preparar los nuevos detalles del asiento contable
-        const journalEntryDetails = details.map(({ ledgerAccount, debit, credit, costCenter }) => ({
+        const journalEntryDetails = details.map(({ account, debit, credit, cost_center }) => ({
             tenant,
-            ledgerAccount,
+            account,
             debit: debit || 0,
             credit: credit || 0,
-            costCenter: costCenter || null,
-            journalEntry: journalEntry._id,
+            cost_center: cost_center || null,
+            journal_entry: journalEntry._id,
         }));
 
         // Insertar los nuevos detalles en paralelo con `insertMany`
@@ -159,11 +175,25 @@ const deleteJournalEntry = async (req, res) => {
         const { _id } = req.params;
         const { tenant } = req;
 
+        const journalEntry = await JournalEntry.findOne({ _id, tenant });
+
+        if (!journalEntry) {
+            return res.status(400).json({ status: 'error', message: 'Asiento contable no encontrado', data: false });
+        }
+
+        const periodoIsOpen = await AccountingPeriod.findOne({ tenant, _id: journalEntry.period, status: 'open' });
+        if (!periodoIsOpen) {
+            return res.status(400).json({ status: 'error', message: 'El periodo contable no existe o está cerrado', data: false });
+        }
+
+        // elimina detalle del asiento contable
+        await JournalEntryDetail.deleteMany({ journal_entry: _id, tenant });
+        // elimina el asiento contable
         await JournalEntry.findByIdAndDelete({ _id, tenant });
 
-        res.status(200).json({ status: 'ok', message: 'Journal Entry deleted successfully', data: null });
+        res.status(200).json({ status: 'ok', message: 'Journal Entry deleted successfully', data: true });
     } catch (err) {
-        res.status(400).json({ status: 'error', message: `Error al eliminar el asiento contable: ${err.message}`, data: null });
+        res.status(400).json({ status: 'error', message: `Error al eliminar el asiento contable: ${err.message}`, data: false });
     }
 }
